Rename build-stage logger module to better indicate its use
