+++ packages/vite-bundler/build.ts
@@ -1,7 +1,7 @@
 import path from 'node:path';
 import fs from 'fs-extra';
 import { cwd } from './workers';
-import Logger from './utility/Logger';
+import Logger from './utility/BuildLogger';
 import Compiler, { BUNDLE_FILE_EXTENSION } from './plugin/Compiler';
 import { Meteor } from 'meteor/meteor';
 import { getBuildConfig, posixPath } from './utility/Helpers';
